Update README.md
Added just a little bit of language to explain the high value from the WCF into the very first section. 

I would have liked to have unified our use of WCT versus WCF as the abbreviation for this, but I can see that there is lots of use of WCT in the code, so opted not to make that change until it had been discussed and agreed with Liwen.

Signed-off-by: Vincent Nunan <vincent.nunan@waivern.com>